pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project]
name = "blog-manager"
version = "0.1.0"
description = "Multi-agent blog management system using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.7",
    "tavily-python>=0.5.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
]

[project.scripts]
blog-manager = "blog_manager.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "ty>=0.0.1a0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["._*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ty]
python-version = "3.10"
